Carefully find min/max coordinates for Azure outputs
