+++ libs/services/google_vision.py
@@ -2,23 +2,7 @@ from google.cloud import vision_v1
 from google.oauth2 import service_account
 
 from libs.region import Rectangle
-
-
-def extract_corners(points):
-    """Identify bounding box for given polygon
-
-    Args:
-        points (list): list of coordinates
-
-    Returns:
-        list: top-left and bottom-right coordinates
-    """
-    min_x = min(points, key=lambda pt: pt[0])[0]
-    min_y = min(points, key=lambda pt: pt[1])[1]
-    max_x = max(points, key=lambda pt: pt[0])[0]
-    max_y = max(points, key=lambda pt: pt[1])[1]
-    
-    return (min_x, min_y), (max_x, max_y)
+from libs.services.utils import extract_corners
 
 
 class GoogleVision:
